🔨 restructure clair parser (#9660)
* :hammer: restructure clair parser

* refactor clair

* refactor clairklar

* update

* flake8
